[sandbox] Turn ArrayBufferExtension into a EPT::ManagedResource
The lifetime of an (out-of-heap) ArrayBufferExtension is tied to the
lifetime of an (in-heap) ArrayBuffer. As it is referenced through the
external pointer table, it should therefore become a ManagedResource to
explicitly tie the lifetime of the extension object to the EPT entry for
it and thereby avoid potentially dangling EPT entries.

Bug: chromium:327550517
Change-Id: I217fcc1d719dcdcc2d9893c72fba616c78eb5b83
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5434873
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93254}