[sandbox] Fix handling of EPT::ManagedResources during table sweeping
With the new managed resources, the external pointer table now
effectively owns certain external objects. As such, when the entry
referencing such an object is freed, the external object must be
invalidated. Previously, we forgot to do that when taking a shortcut
during sweeping: if an entire segment of the table was empty, we simply
deallocated that segment but forgot to check for managed resources. This
CL fixes that.

Bug: chromium:333493383
Change-Id: I44bbd409e8eff7fa5aea0d1c092eab6c1a353665
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5438117
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: Samuel Groß <saelo@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93257}