[flags] Extract name from complex implication
Previously, in DEBUG builds, FindImplicationByName could fail in the
event that an implication involving comparison operators is used,
such as "--max-opt 1". The implication "max_opt < 2" would fail to
match with the correct Flag, and a DCHECK asserting that the found
flag name matches the implication flag would fail.

This CL fixes the issue by allowing for suffixes such as the above,
when looking up flag implications. It works correctly as long as the
flag with the suffix compares against other flags exactly as the flag
without the suffix would compare. This is necessarily true if the
suffix in implications begins with a space, e.g., "max_opt < 2", which
is exactly what currently happens in our flag definitions.

Change-Id: I310945ca283a6dce241c5626a0a1cc2d86dc9243
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5438116
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93294}