[string] Refactor SHA256 script content hashing
This CL changes how we calculate the SHA256 hash of a V8 string,
currently used for the CallSite API and the inspector to get script
content hashes.

The current approach requires allocating a copy of the string
(via String::ToCString) and hashing that. Instead, we can iterate
the string ourselves and build the hash as we go. In microbenchmarks
this is slightly faster and doesn't require an extra copy of the
string.

Note that using SHA256_update on one-byte-non-cons string chars
directly didn't improve performance so we don't gain anything by
optimizing further based on string representation.

Drive-by: Add a new debug-interface method so both code paths that
calculate script hashes use the same approach.

Bug: 326554286
Change-Id: Iebbd501a67f164822c2d2d8f3e24863e3477ec40
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5423429
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93317}