Add .size for PushAllRegistersAndIterateStack
When profiling function entry events on Android, the orderfile generator
often records an offset within PushAllRegistersAndIterateStack, but
cannot attribute the offset to this function because of the zero size
associated with the ELF symbol. This hits a rare assertion in the
orderfile pipeline. If we remove the assert, the symbol will not be in
the orderfile, which is not ideal for this potentially hot code.

Add the size using the `.size` directive to the assembler.

Limit the change to the ISAs where Android orderfile is supported or
going to be supported in overseeable future, including emulators.
Compile out the directive for __APPLE__ and Windows.

Bug: chromium:330761384
Change-Id: Ie175bf676942fa9af798d74b8cf381977db31a3c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5443265
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93319}