[wasm] Allocate code spaces close to each other
If we call between different Wasm modules we get better performance if
the code spaces are allocated close to each other. We thus remember the
end pointer of the previously allocated code space in the
WasmCodeManager (global in the process), and use it as a hint for the
next code space allocation.

This mechanism replaces the explicit hint we were passing for allocating
an additional code space within one module.

R=jkummerow@chromium.org
CC=dlehmann@chromium.org

Change-Id: If8d8056a3626e4c1e918e29e5a178b72498c2d37
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5443266
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93298}