[cleanup] Refactor header files for string builders
So far, string builders were defined in string-builder-inl.h and
implemented in string-builder.cc, while there was no string-builder.h
file. This CL splits the header file in two parts, as is usual. It also
cleans up the implementation of CallPrinter, which was unnecessarily
complicated because of the rule about not including -inl.h files in .h.

Change-Id: I2d718d510ead43f7811b5efc3a005c35ad9b7e7c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5443843
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93382}