[deoptimizer] Untangle frame translation iterator from heap
The DeoptimizationFrameTranslation::Iterator requires the byte array to
be allocated on the heap. As wasm is isolate-independent, it cannot
have this data on the heap and it cannot create a "heapified" copy on
demand (as it isn't safe to allocate on the heap during frame
translation).

This CL moves the implementation of the iterator to a class taking a
raw byte vector. For JS, the original class is a trivial wrapper around
this implementation.

Bug: v8:14667
Change-Id: Ic8bcbe3466914c725e5d6b638d91052be82b37d5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5443269
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93337}