Reland "[sandbox] Turn ArrayBufferExtension into a EPT::ManagedResource"
The lifetime of an (out-of-heap) ArrayBufferExtension is tied to the
lifetime of an (in-heap) ArrayBuffer. As it is referenced through the
external pointer table, it should therefore become a ManagedResource to
explicitly tie the lifetime of the extension object to the EPT entry for
it and thereby avoid potentially dangling EPT entries.

Bug: chromium:327550517
Change-Id: Id9fe16fafb211e175ba2ff6ce6d8e922bfbf813f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5441118
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93341}