[profiler] Use gc_state to detect GC in GetStackSample
With the LOGGING VMState, we might overwrite the GC VMState (if e.g. we
are logging a move event). This is a problem in the profiler, which was
avoiding doing deep analysis of samples in GC, since the pointers it
would look at could be invalid during evacuation.

Use the heap's recorded gc_state in addition, which is more precise than
the VMState and won't be clobbered by VMState changes. We still need the
VMState check though, since some parts of GC (e.g. heap verification)
are outside the "in-GC" state.

Bug: v8:14713
Change-Id: I756cc8f4d402effc12492b5a3c436bd2ba111a3e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5439416
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93358}