[wasm][fastcall] Handle exceptions thrown in fast API calls
An API function called with a fast API call can throw an exception by
storing the exception on the isolate and setting a flag on the isolate.
Generated code then checks if the flag on the isolate is set, and if so,
call to the runtime to trigger stack unwinding through the normal
exception handling of the C-Entry stub.

The implementation for throwing an exception is in
https://crrev.com/c/5442855, this CL adds the wasm implementation for
catching the exception.

R=thibaudm@chromium.org

Bug: 41492790
Change-Id: I380a3597001bfb073a037689ccd9a8ba603829f4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5445546
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93360}