[wasm] Use immutable loads on some protected fields
Protected fields currently didn't have support for marking the loads as
immutable. This CL adds that to Turbofan and Turboshaft and marks all
relevant field loads from the instance as immutable.

Drive-by: Remove a TODO which turns out to be inactionable (and is not
faster as bitwise or because it cannot be folded into a memory operand).

R=ahaas@chromium.org, dmercadier@chromium.org

Bug: chromium:332120965, v8:14564
Change-Id: Ied41a940b460ae29d294bd96e42231c8d5fdbdfd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5441121
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93366}