[fastcall] Support throwing exceptions
A function called with a fast call can set a flag on the isolate to
indicate that an exception has been thrown. As for normally throwing an
exception, the exception has to be stored on the isolate anyways.
Generated code then checks after a fast call if the flag on the isolate
is set. If so, a runtime function is called that triggers stack
unwinding through the normal exception handling of the C-Entry stub.

Additionally deopt info is added to fast API calls, not because deopt is
possible, but because deopt info is needed to generate stack traces for
exceptions.

The CL does the following changes:
(1) The TurboFan node `FastApiCall` gets two control outputs for
`IfSuccess` and `IfException`;
(2) The Turboshaft operation `FastApiCallOp` gets the ability to throw
an exception, and gets a FrameState input so that deopt information is
generated for the call;
(3) The instruction selector adds to a `kArchCallCFunction` the
information if deopt info is needed for the call, and the code generator
generates deopt info in that case;
(4) A new flag is added to the isolate that allows generated code to
check after a fast API call whether an exception has been thrown.

R=dmercadier@chromium.org

Bug: 41492790
Change-Id: If283205d66873fd5b7ac24462f950e36d1e52fef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5442855
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93357}