[wasm] Add metadata size to memory consumption estimate
This introduces a WasmCode::EstimateCurrentMemoryConsumption method to
account for the size of the WasmCode objects themselves plus the
metadata which is allocated on the C++ heap per code object.

Note that the previous `ContentSize(owned_code_)` and
`ContentSize(new_owned_code_)` were only accounting for the size of the
`unique_ptr`, not the size of the `WasmCode` objects.

R=jkummerow@chromium.org

Change-Id: I182a19b9d81892ea0366067ba936b23695eb1a45
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5453413
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93368}