[heap] Fix deadline check in ProcessMarkingWorklist
The objects_processed counter is used to avoid checking TimeTicks::Now
too often. The check was using bitwise-and instead of modulo, which
means that instead of checking every 128 objects, it was checking after
each object for the first 127 objects, then wasn't checking for the next
128 objects, then again checking after each object for the next 128
objects, and so on.

Change-Id: Iaae0c01716ec50c85c9fb940864d41547a5f2501
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5458151
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Auto-Submit: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93397}