[snapshot] Disable ICF in mksnapshot
Identical code folding can cause external references to be de-duplicated
(e.g. merging memcpy and memmove). This is ok, as long as the same
functions are folded both in the serializing and deserializing binaries.
If the serializing binary merges two functions that the deserializing
binary didn't, then the snapshot deserialization will get the wrong
function reference. This can happen due to whole program optimization of
the snapshot serializing binary.

We can prevent this by disabling ICF for snapshot generation, ensuring
that there are no duplicate references when generating the snapshot.
It's ok for the mismatch to be in the opposite direction, i.e. for the
deserializing binary to have merged functions that the serializing
binary didn't, so this is a safe operation and doesn't affect the
shipped binary.

This also adds a small section to the start of the startup snapshot
which lists all duplicate external reference table entries. The
deserialization verifies that duplicates stayed duplicated. This allows
us to support some duplication in the serializing binary as long as
there is matching duplication in the deserializing binary -- this will
be the case, for example, if the serializing and deserializing binary is
the same (e.g. in tests, or in node with --binary-snapshot).

Bug: chromium:41321762, chromiun:327245456
Change-Id: If19584d19307bec5ea4d8b6a511151593d101988
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5447267
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93396}