[fastcall] Check exception directly instead of exception flag
The initial implementation of exception handling of fast api calls
introduced a flag in the isolate data that indicates if an exception was
thrown in the API function. However, it seems possible to just check the
exception field in the isolate data directly. If the exception field
contains the_hole_value, then there is no exception. Otherwise an
exception has been thrown.

Bug: 41492790
Change-Id: I8dcda42c123f195b7c40a34022c25b4361ad407c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5458172
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93410}