[ic] Store the setter in the FBV
For accessors, instead of storing the descriptor index + holder in the
StoreHandler, store the setter directly (avoiding the
map->descriptor->pair->setter hops). For the non-prototype case, where
there's no StoreHandler, store the AccessorPair directly as a weak
handler instead of the Smi handler. We can't store the setter here
directly, because it could be in new space, and then we can't use it in
the stub cache.

This is a follow-up for http://crrev.com/c/3865557.

Change-Id: I96e0cb4f55c2745a968e8e54e2e674132eaea770
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5458514
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93431}