heap: Use fast check for static RO objects
On builds with static roots we can filter out RO objects based on a
compressed value (Tagged_t). This conveniently also filters out
"small" Smis in the same range.

The check is not complete as not all RO objects are static roots and
as such the page-level checks cannot be removed.

During Scavenge on S3 the simple check filters out 40% of objects as
static RO roots and 17% as Smis.

Change-Id: Ibc180727b7aedfc46301b4c11930619d1f226ea3
Bug: chromium:335479495
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5459151
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93430}