[regexp] Fix memory leak in Experimental
When discarding threads, they are now correctly destroyed and the
underlying memory is given back to the allocator.

Additionally, clearing the threads list was not done properly, using
the `DropAndClear()` function, instead of `Rewind(0)`. This caused the
underlying pointer to be dropped without being freed, while the
desired behavior was to empty the list without losing the allocated
space.

A test was added, for which the memory consumption used to explode, and
is now constant in the size of the input string.

This work was supervised by Aurèle Barrière, from SYSTEMF at EPFL.


Bug: v8:14691
Change-Id: I1ccbded25b8b711c4c9fde46c483ed7fb9be7859
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5454791
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93459}