Speed up heap snapshots
When producing a heap snapshot, V8 starts by counting the objects in the
heap so that it can report its progress. It currently does liveness
analysis during this phase, which is unnecessary because it had just
collected garbage, so the number of dead objects should be low, and a
precise result is not required. By skipping the liveness analysis, we
can save over 10% of the snapshot generation time (1.7 seconds out of 13
on an 80 MB Google Docs page).

Bug: 42203857
Change-Id: I63411a4154b0c1fb5980feafdd5176b43f246128
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5466804
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93658}