[handles] Keep list of blocks with young nodes
This CL changes the way we track young TracedNodes in the GC. So far
we tracked all such nodes in the young_nodes_ vector. Appending to the
vector is a large part of TracedNode creation cost.
Instead of tracking individual nodes we now track blocks instead.
This reduces the CPU cycles spent in TracedNodes::Create by about 66%
in Speedometer 3. However, this also means that young nodes iteration
in the MinorGC becomes more expensive as we now need to iterate the
full block. Fortunately this neither regressed total CPU cycles in
minor GC nor the total score.

As a side-effect of this CL we are now able to get rid of
empty_block_candidates_ as we are now able to quickly remove
individual blocks from the young_blocks_ free list. With young_nodes_
we couldn't do this as we didn't know where a TracedNode was stored in
the vector without iterating the whole vector.

Bug: 335695521
Change-Id: I4418474aa078953636f276b4fad2b13e8559de66
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5447268
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93483}