[liftoff] Carve out VarState into own header
This CL moves the VarState out of the liftoff assembler.
VarState describes the location of a wasm variable (like a value on the
wasm value stack, a local or a parameter) and defines its "physical"
location (i.e. a register, a stack slot or a constant).

This information is needed by the deoptimizer to be able to construct a
Liftoff frame as, differently to unoptimized JS frames, the wasm frame
layout depends on internal liftoff details like register allocation.

Bug: 42204618
Change-Id: I5b55459541fd54796bcb4bbab0ef854fefbda27f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5463596
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93466}