Restore microtask CPED before the early return
MicrotaskQueue::RunMicrotasks was recently updated to better handle
CPED. Unfortunately, if the microtasks were interrupted using
TerminateExecution, the captured CPED would not be properly restored.

This ended up causing us a bit of pain this week when updating to 12.4. In certain cases, TerminateExecution was called while the microtasks were processing which led to the wrong context being propagated in a few spots, leading to a couple of difficult to track bugs. Moving the CPED restoration before the early return fixed our issues.

Change-Id: Ife311d1de55a34a531f4c905b5107ad8c78935f8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5468303
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93495}