[turboshaft][wasm] Fix wrong FrameState input order and count
The wasm stack values were pushed in the "wrong" order (i.e. not
consistent with the order of the VarStates in liftoff.

There was also a mistake on the count of locals.
The correct calculation is:
- values on the value stack `decoder->stack_size()`
- wasm function locals `ssa_env_.size() - param_count`
Unfortunately, the decoder value stack already describes the state
after the call_ref instruction, so we need to adapt the following:
- add the call_ref callee
- add the wasm trusted instance (implicit argument to the call_ref)
- add the call's explicit arguments
- subtract the call's results (as these are pushed after the call)

Drive-by change: Include deopt data size in WasmCode to wasm memory
usage estimation.

Bug: 42204618
Change-Id: Ic460f47b27dcfdaaa2054bcebcb4f7eae72d19bc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5468102
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93475}