[string] Fix OOB table read in ToLatin1Lower
ToLatin1Lower looks up a uint16_t value in a 256 valued array, under the
assumption that the uint16_t value is actually a one-byte character that
happens to be stored in a two-byte string (e.g. a one-byte SlicedString
with a two-byte parent due to externalization). Ensure that this table
lookup can't go out of bounds under heap corruption, by changing the
function to take a uint8_t and doing a DCHECKed truncate of the uint16_t
character to uint8_t.

As a drive-by, audit the surrounding code (e.g. ToLatin1Upper), and add
similar guards -- these won't do a similar table lookup currently, but
maybe they would in the future, and it's good to anyway be consistent
and explicit about this.

The regression test requires support for the memory corruption API in
mjsunit (in particular, to skip tests if the API isn't there), so add
that too.

Fixed: chromium:335810507
Change-Id: I0e5d3ca31dda3bd4ce4dc03c3a2a802290489f2d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5471973
Reviewed-by: Samuel Groß <saelo@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93502}