[wasm-stringref] Make stringviews non-nullable
The previous in-between state where views were supertypes of 'none'
but not subtypes of 'any' was too inconsistent to be workable. This
patch forces the three view types to be non-nullable references,
which effectively puts them into single-occupant type hierarchies
of their own.

Fixed: 335640704
Bug: 42202811
Change-Id: I58f9a590c85bf90023d2b0bcdf1b1fbd071cda99
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5471674
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93534}