[wasm][revec] Fix Load/Store reduction when stores are reversed in seed
Current reducing Load/Store will always take the address from the first
node in PackNode to generate the revectorized node. But if the nodes
are in reversed order in PackNode, it's possible that ChangeOp of the
first load's index is not emitted yet and cause compile time crash.

This patch will fix this issue by taking the index from the current
visiting node and only getting the offset from the first node.

Bug: v8:12716
Change-Id: I8ca3715b668d1bbcfc6fc2be32f3685c201c289a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5480037
Commit-Queue: Yolanda Chen <yolanda.chen@intel.com>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93545}