[wasm] Fix Liftoff slot count for exception handling
The frame slot count reported by Liftoff is not always exact: it
includes the registers pushed on top of the stack in out-of-line code.
So if we are currently in a regular (not out-of-line) call, then the
actual slot count is smaller.

This is good enough for the purposes of the GC, which will simply
ignore the upper slots that are not marked as references.

However, the unwinder uses the same value to compute the stack pointer
from the frame pointer, and this one should be exact. Otherwise the SP
in the landing pad is not going to match the actual frame size.

Therefore register the spill count for OOL code in the WasmCode and
use it to compute the exact frame size in the unwinder. This assumes
that we don't handle exceptions in out-of-line code.

R=ahaas@chromium.org

Bug: 336214779
Change-Id: Ib793c02c5958f9e0da547da48aeec2bf86296273
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5484026
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93573}