[sandbox] Introduce Sandbox.getFieldOffsetOf API
This API exposes a pre-defined list of field offsets (essentially
the FooObject::kBarOffset constants) to JavaScript via the memory
corruption API. This can be useful for example for regression tests so
that these do not need to hardcode field offsets (which may change).
Furthermore, when a field is deleted, any tests using it will
automatically fail (as the API will then throw an exception) and can
then be modified or deleted.

For now, the list of supported fields is manually maintained. In the
future, it would be great to somehow automatically generate that list
based on the C++ field offsets.

Bug: chromium:42202821
Change-Id: I8000fa61dcc3bcafd3c48d7f4429bd5bcc4631e0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5487939
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93576}