[maglev][osr] Fix context slot analysis confusion
When compiling OSR code an inlinee's context can alias with the current
context at the OSR entrypoint. The reason is that the closure
creation happens outside of the part of the function that is being
compiled. Therefore, we see the same context twice -- once as the
InitialValue of the context register and once as a constant originating
from the inlinees context slot. For now we make the analysis a bit more
conservative.

Fixed: chromium:41494766
Change-Id: I9a410fc72775df9ee12afa1676047374d9640bd1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5481577
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93580}