[profiler] Increase sample processor thread stack size to 256KB
This CL fixes an issue where the sample processor thread failed to
start due to insufficient stack size. Concretely pthread_create bailed
with an "Invalid argument" errno.

pthread_create checks for a couple of things when determining if the
provided stack size is large enough. One of them is the size of the
static TLS area which is determined by the dynamic loader:

https://elixir.bootlin.com/glibc/latest/source/nptl/allocatestack.c#L351

When V8 runs under the Centipede fuzzer + ASAN, the static TLS size
alone is 96KB, so the currently 64KB are not enough. Most of it is
probably due to Centipede: They intercept pthread_create and allocate
per thread machinery:

https://source.chromium.org/chromium/chromium/src/+/main:third_party/fuzztest/src/centipede/runner_interceptors.cc;l=56;drc=e6eb1a564fd608d3ee98a7010ee3120e863bef03

Drive-by: CHECK if the sample process thread actually started,
otherwise we'll silently pretend it's running and only later fail
when we try to pthread_join a thread that never ran.

Fixed: 335342168
Change-Id: I57dc89a55e807d7e4471a2f63327a2d574439b18
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5490091
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93609}