cppgc: Prefer sweeping in idle tasks
This patch changes the strategy used for cppgc (Oilpan) sweeping.
Until now we preferred finishing sweeping as soon as possible using
regular main trhead tasks and also reusing memory as much as possible
on allocation. This apporach essentially minimizes RSS.

This change moves us in the tradeoff space towards a strategy where we
try to get out of the way of the mutator (e.g. main renderer) as much
as possible at the expense of possibly using slightly more memory.

In particular this change:
a. Uses idle tasks to sweep memory and run finalizers as much as
   possible. We still preserve a fallback in case idle tasks are not
   enabled or there's no idle time. The fallback is ensured via
   running a delayed task that checks for progress.
b. Avoids synchronously sweeping on allocation when we are further
   away from the next GC (via checking the allocation limit).
   Synchronous sweeping is enabled when getting closer to the actual
   heap limit to split up sweeping work and avoiding a huge pause at
   the start of marking.

Change-Id: Ie516c67c9a2151d150b65f82229dba8aa754b43c
Bug: chromium:333981063
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5454401
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93612}