[profiler] Fix DCHECK failure when snapshotting WasmInstanceObject
When generating heap snapshots, ExtractWasmInstanceObjectReference()
was adding references from the corresponding WasmTrustedInstanceData
to WasmInstanceObject. However, this was violating the invariant that
field offsets fit into the current object.

This CL solves this by annotating these references only on the
WasmTrustedInstanceData object but not for the WasmInstanceObject
anymore.

It also tries to prevent such violations in the future by adding a
max_pointers_ field and DCHECKing that the field offset/index is not
too large. This DCHECK should fail more predictably and while still
handling the broken object.

Bug: 42204595
Change-Id: Ia21a7ba2f029390a493164145c8324c64e32083e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5492423
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93617}