[base] Optimize VLQEncodeUnsigned
Change VLQEncodeUnsigned from a loop that adds continuation bits, into
straight code writing the five bytes, with gotos jumping to the
appropriate number of bytes to write based on the value.

In particular, this avoids writing to the same location twice (once for
the value bits, once for the continuation bit), which also means we can
simplify the API to not require process_byte to return a pointer to the
last written byte.

This appears to be ~2x faster than the loop:
https://quick-bench.com/q/TPfnCwLsgDQz4hvUwdJk-pHPwVU

Change-Id: Ia732c0103a140ec5c72ea558dbd71952e98260df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5496956
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93624}