[heap] Optimize sweeping through tasks
When sweeping in tasks we generally observe that concurrent sweeping
is not done when starting sweeping on the main thread. However, by the
time main thread sweeping is done, the concurrent sweeper is also
almost always done. We can use the remaining time to avoid scheduling
another task and eagerly finish sweeping in this case.

Drive-by: Fix IsConcurrentSweepingDone() for cases where sweeping is
fully finished (in which case the handle would be set but not valid).

Bug: chromium:333981063
Change-Id: Ic35226f915b740c76952bed9d1a6bb4278258ecb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5498376
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93648}