[heap][sandbox] Add young generation for ExternalPointerTable
When the sandbox is enabled, instead of allocating external pointer table
entries in one big space that is only collected as part of a major GC,
add a spatially partitioned young generation.  After a minor collection,
eagerly sweep that new space, so it is ready for fresh allocations.
When objects are promoted, their EPT entries get evacuated to the old
space.

Design doc:
https://docs.google.com/document/d/1Pvr4RbG_ZiaYZ-VkE22bSyCOiOntpbSBNJIhv45JTKk

Bug: chromium:40643874
Change-Id: I08978c087c809328fd614325c9b238e4e5a0994a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5185345
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Andy Wingo <wingo@igalia.com>
Cr-Commit-Position: refs/heads/main@{#93653}