Set up proper installation guard for JSPI.
Since JSPI & type reflection can be installed after user code has
executed, we need to make sure that (a) it can only be used if
properly installed, and (b) anytime JSPI is enabled, the installation is also completed.

Created a boolean flag on the context object to make sure that JSPI features cannot be accessed unless it is properly installed.

Bug: 336358915
Change-Id: I6d95169c325257d963771b5471e763f76fcd2ac9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5484832
Reviewed-by: Adam Klein <adamk@chromium.org>
Commit-Queue: Francis McCabe <fgm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93664}