[tests] Stop using deprecated v8::ObjectTemplate::SetAccessor(..)
... for interceptor tests.
The getter/setter callbacks were used for checking whether the get/set
request was correctly handled by interceptor in a child object.

The usual replacement SetNativeDataProperty(..) doesn't fit here
because according to JavaScript spec the setter callback should not be
called in case a property is stored to a child object.
So, the tests are updated to use JavaScript accessor callbacks instead.

Bug: 336325111
Change-Id: I8ac6f160635cf1fbbbf90e7037bea18f076cc630
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5498423
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93674}