[maglev] Properly fix context slot analysis confusion
Aliasing between contexts can also happen without OSR, in case both the
initial context as well as the constant context of an inlinee refer to
the same runtime context. Thus the initial fix in crrev.com/c/5481577
was not complete.

To properly solve the issue we introdue a may-alias analysis between
contexts based on scope infos. The analysis has two parts:

1. Keep track if we loaded both from a constant as well as a dynamic
   constext in `KnownNodeAspects::may_have_aliasing_contexts`.
2. In case we did, on every store check all the cached values for other
   contexts which could refer to the same ScopeInfo and clear them.

To acchieve (2) we remember the association between context SSA values
and ScopeInfo's where neccessary on the MaglevGraph.

In the case of OSR it can happen that we cannot attribute the scope info
at the OSR entrypoint with certainty. In this case we assume that it can
alias with any scope.

Fixed: chromium:41494766
Change-Id: I1227446cd29cc758ce90738217da337c7a20a52a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5497160
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93683}