[wasm][jspi] Fix IsOnCentralStack check
When switching to a target stack, we should look at the SP, not the
stack base, to determine if the target stack is on the central stack or
not.
This makes a difference if the target stack is originally a secondary
stack, but is in the middle of a JS or C++ call. In this case, the
stack memory still holds a pointer to the secondary stack memory, but
the SP is temporarily inside the central stack.

Drive-by: remove obsolete comment.

R=jkummerow@chromium.org

Bug: 338122900
Change-Id: I191df8f37acaa2b21e12f0a5a76d783554eb7e88
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5515167
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93698}