[deoptimizer][wasm] Add support for parameter stack slots
This adds support for deopting a function that has parameters passed
via stack slots.

This requires the deoptimizer to include the parameter stack slots
into the FrameDescription which also requires to include the stack
objects between the frame and the parameter slots (2 slots) and then
to "copy" this over to the output FrameDescription.

Bug: 42204618
Change-Id: I74884ca77a4817ccb7b0f7319bd81fdb9014b5a1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5498887
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93741}