[heap] Don't visit embedder fields when not supported
https://crrev.com/c/5515168 allowed for passing a WrapperDescriptor in
a disabled state to enable deprecation of GC support in embedder
fields. The disabled descriptor is never checked so far though which
leads to OOB in embedder fields when using a descriptor that should
not be used for tracing.

Bug: 338411141
Change-Id: Ia15840c4221d0a3b93eab04f6f8f251825e0ce90
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5518762
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93763}