[etw] Set ETW handler only after read only heap is sealed
When ETW logger starts, it iterates heap objects to find all the code
objects. But, if the read-only heap is not ready, it iterates over the
limit, which is not initialized with fillers yet.

The crash can be reproduced by the following.
cctest test-serialize/ContextSerializerCustomContext --enable-etw-stack-walking

Change-Id: I99aa35212a4d7f02d488b7ad9730ba4f6764f699
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5523693
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Choongwoo Han <choongwoo.han@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#93800}