[deoptimizer][wasm] Support Simd128 values
Attention: This CL changes how JavaScript handles double registers!

This CL adds support for Simd128 values in the deoptimizer for wasm.
Note that wasm deoptimizations are WIP and the current state is only
implemented for x64.
However, to support Simd128, the deoptimizer needs to know about
Simd128 register values which overlap with double register values.
So we need to migrate double register value handling to use Simd128
registers instead (meaning the 128 bit registers instead of the 64 bit
variants).

For JavaScript, this means that we need to move twice the amount of
data for the double registers. In most cases: Push the value from the
register to the stack, then copy it from the stack to the input
FrameDescription, copy it from the FrameDescription to the
TranslatedValue, from there into the outgoing FrameDescription and
finally from there back into the register (if it's still in a register
that is).

Bug: 42204618
Change-Id: I7253d037460a095f893cea239d448b1977007d41
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5513446
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93911}