[shared-struct] Make lockAsync fail instead of crashing
It is possible for the call to the promise_then builtin (that the
asyncLock uses to queue internal promises) to fail. E.g. reaching max
stack size.

This CL allows the lockAsync and waitAsync APIs to fail instead of
crashing in a CHECK.

Bug: 42202493
Change-Id: Ie8a3150e96ff84aefa4bfc77f07913825bea221f
Fixed: 339043696
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5528989
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Commit-Queue: Luis Pardo <lpardosixtos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#93834}