[cppgc] Fix registering on-stack movable slots
We assume that there's no on-stack slots for movable references as
compaction doesn't run when finalizing GC with stack. Unfortunately,
it is possible to trigger write barrier for on-stack slots. It's
expensive to filter those out generically (as that would require an
on-stack check for each slot) and marking does work with such slots.
As a result, filter them specifically before storing movable
references.

Bug: 339967265
Change-Id: I2449a47c4df92a4d117789ce7b93e6ef8075dc93
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5534767
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93858}