[sandbox] Check array initial map was not corrupted
DependOnInitialMapInstanceSizePrediction uses the default initial
map for the array_function to calculate the slack size.

Later we update the initial map depending on the elements kind.
We assume that the slack calculated by this updated initial map
is the same.

So in the CL we add a check in CapturedObject::set to be sure
we don't overflow the CapturedObject.

Drive-by: We refactor ReduceArrayConstructor to make the
initial map logic more explicit.

Fixed: 339517309
Change-Id: Ib7de77738895c1ff2ccfced2d4bc5499c18a5ca2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5534744
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93860}