[heap] Fix allocation tracker line end retrieval
Tests in NodeJS with DCHECK enabled were failing because of two
different problems:
- One was that we were also disallowing heap allocations in the
serialization stage. But NodeJS tests process the heap snapshot
result in JS, so all those were broken.
- But, also, the code that would retrieve from line ends would assume
all the scripts populated line ends in the snapshot. But this was
wrong. To fix it, this patch adds another storage of the line ends
in the allocation tracker. This storage needs to keep weak references
to the scripts so we do not leak the line ends data when scripts are
disposed.

This change keeps a weak reference to the scripts so, when they are
freed, the line ends cache is also freed as it is not useful anymore.

Node issue: https://github.com/nodejs/node-v8/issues/282

Change-Id: I4314d707903175f0005893e9aa06d3ae52fc57f8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5531355
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94418}