Shadow Test262 staging tests with V8's staging tests
Patch #1 contains a local staging implementation of a test262 staging test. The test gets to run twice (once for every locations), but it uses the local staging implementation for both runs. Things break on resultdb as we generate the same key for both runs. In the CQ for patch #1 we error out:
 - task https://chromium-swarm.appspot.com/task?id=6966b3197f9bf810 with key test262/staging/set-is-subset-table-transition.VariantProc-default-0.Rerun-1
 - task https://chromium-swarm.appspot.com/task?id=6966b31d4d1d8110 with test262/staging/set-is-subset-table-transition.VariantProc-default-strict-0.Rerun-1

Patch #2-13 implements and tests the shadowing mechanism where we exclude test262 staging tests that have an equivalent V8 staging test. CQ passes.

Patch #13 runs with a shadowing test that fails.

Patch #14 removes the shadowing test.


Bug: 339207824
Change-Id: Ie299f5d9e364fb396801426f629840ca73b8ce14
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5518172
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93893}